refactor(node:stream): split node_stream.rs into topical sub-modules (#1987)
node_stream.rs had grown to ~5,980 lines during the node:stream parity wave
and was allowlisted out of the 2,000-line file-size gate. Split the monolith
into topical sibling modules and drop both node:stream allowlist entries.

Pure code-motion. The parent keeps the shared vocabulary (tag constants,
hidden-key constants, closure/microtask stubs, js_node_stream_method_* FFI
getters, JSON hooks) and re-globs each child with pub use so the pub extern
"C" FFI entry points stay reachable at perry_runtime::node_stream::* for
other crates; each child opens with use super::*. New files:
node_stream_{keys,pipeline,iter_helpers,dispatch,readwrite,constructors}.rs.
Every file is now under 2,000 lines (largest: readwrite ~1,865).

Validated: full runtime/stdlib/perry release build, node_stream unit suite
(the 4 readable-default-read-error failures are pre-existing on main, broken
by #2441 independently of this code-motion), and a cross-category node-suite
stream parity sample byte-identical to Node.
